+++ printing/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 5.3.1
+- Added a boolean toggle to show/hide debug switch
+
 ## 5.3.0
 
 - Fix raster crash on all OS.